Move geometry to separate folder
