+++ seismiqb/src/geometry/segy.py
@@ -1,0 +1,545 @@
+import os
+import re
+import sys
+import shutil
+import itertools
+
+from textwrap import dedent
+from random import random
+from itertools import product
+from tqdm.auto import tqdm
+
+import numpy as np
+import pandas as pd
+import h5py
+import segyio
+import cv2
+
+from ..utils import find_min_max, file_print, compute_attribute, make_axis_grid, \
+                   fill_defaults, parse_axis, get_environ_flag
+from ..utility_classes import lru_cache, SafeIO
+from ..plotters import plot_image
+
+
+from .base import SeismicGeometry
+
+
+
+
+class SeismicGeometrySEGY(SeismicGeometry):
+    """ Class to infer information about SEG-Y cubes and provide convenient methods of working with them.
+    A wrapper around `segyio` to provide higher-level API.
+
+    In order to initialize instance, one must supply `path`, `headers` and `index`:
+        - `path` is a location of SEG-Y file
+        - `headers` is a sequence of trace headers to infer from the file
+        - `index_headers` is a subset of `headers` that is used as trace (unique) identifier:
+          for example, `INLINE_3D` and `CROSSLINE_3D` has a one-to-one correspondance with trace numbers.
+          Another example is `FieldRecord` and `TraceNumber`.
+    Default values of `headers` and `index_headers` are ones for post-stack seismic
+    (with correctly filled `INLINE_3D` and `CROSSLINE_3D` headers),
+    so that post-stack cube can be loaded by providing path only.
+
+    Each instance is basically built around `dataframe` attribute, which describes mapping from
+    indexing headers to trace numbers. It is used to, for example, get all trace indices from a desired `FieldRecord`.
+    `set_index` method can be called to change indexing headers of the dataframe.
+
+    One can add stats to the instance by calling `collect_stats` method, that makes a full pass through
+    the cube in order to analyze distribution of amplitudes. It also collects a number of trace examples
+    into `trace_container` attribute, that can be used for later evaluation of various statistics.
+    """
+    #pylint: disable=attribute-defined-outside-init, too-many-instance-attributes
+    def __init__(self, path, headers=None, index_headers=None, **kwargs):
+        self.structured = False
+        self.dataframe = None
+        self.segyfile = None
+
+        self.headers = headers or self.HEADERS_POST
+        self.index_headers = index_headers or self.INDEX_POST
+
+        super().__init__(path, **kwargs)
+
+
+    # Methods of inferring dataframe and amplitude stats
+    def process(self, collect_stats=False, recollect=False, **kwargs):
+        """ Create dataframe based on `segy` file headers. """
+        # Note that all the `segyio` structure inference is disabled
+        self.segyfile = SafeIO(self.path, opener=segyio.open, mode='r', strict=False, ignore_geometry=True)
+        self.segyfile.mmap()
+
+        self.depth = len(self.segyfile.trace[0])
+        self.delay = self.segyfile.header[0].get(segyio.TraceField.DelayRecordingTime)
+        self.sample_rate = segyio.dt(self.segyfile) / 1000
+
+        # Load all the headers
+        dataframe = {}
+        for column in self.headers:
+            dataframe[column] = self.segyfile.attributes(getattr(segyio.TraceField, column))[slice(None)]
+
+        dataframe = pd.DataFrame(dataframe)
+        dataframe.reset_index(inplace=True)
+        dataframe.rename(columns={'index': 'trace_index'}, inplace=True)
+        self.dataframe = dataframe.set_index(self.index_headers)
+
+        self.add_attributes()
+
+        # Collect stats, if needed and not collected previously
+        path_meta = os.path.splitext(self.path)[0] + '.meta'
+        if os.path.exists(path_meta) and not recollect:
+            self.load_meta()
+        elif collect_stats:
+            self.collect_stats(**kwargs)
+
+        # Create a matrix with ones at fully-zero traces
+        if self.index_headers == self.INDEX_POST and not hasattr(self, 'zero_traces'):
+            try:
+                size = self.depth // 10
+                slc = np.stack([self[:, :, i * size] for i in range(1, 10)], axis=0)
+                self.zero_traces = np.zeros(self.lens, dtype=np.int32)
+                self.zero_traces[np.std(slc, axis=0) == 0] = 1
+            except ValueError: # can't reshape
+                pass
+
+        # Store additional segy info
+        self.segy_path = self.path
+        self.segy_text = [self.segyfile.text[i] for i in range(1 + self.segyfile.ext_headers)]
+        self.add_rotation_matrix()
+
+    def add_attributes(self):
+        """ Infer info about curent index from `dataframe` attribute. """
+        self.index_len = len(self.index_headers)
+        self._zero_trace = np.zeros(self.depth)
+
+        # Unique values in each of the indexing column
+        self.unsorted_uniques = [np.unique(self.dataframe.index.get_level_values(i).values)
+                                 for i in range(self.index_len)]
+        self.uniques = [np.sort(item) for item in self.unsorted_uniques]
+        self.uniques_inversed = [{v: j for j, v in enumerate(self.uniques[i])}
+                                 for i in range(self.index_len)]
+
+        self.byte_no = [getattr(segyio.TraceField, h) for h in self.index_headers]
+        self.offsets = [np.min(item) for item in self.uniques]
+        self.lens = [len(item) for item in self.uniques]
+        self.ranges = [(np.max(item) - np.min(item) + 1) for item in self.uniques]
+
+        self.cube_shape = np.asarray([*self.lens, self.depth])
+
+    def collect_stats(self, spatial=True, bins=25, num_keep=5000, pbar=True, **kwargs):
+        """ Pass through file data to collect stats:
+            - min/max values.
+            - q01/q99 quantiles of amplitudes in the cube.
+            - certain amount of traces are stored to `trace_container` attribute.
+
+        If `spatial` is True, makes an additional pass through the cube to obtain following:
+            - min/max/mean/std for every trace - `min_matrix`, `max_matrix` and so on.
+            - histogram of values for each trace: - `hist_matrix`.
+            - bins for histogram creation: - `bins`.
+
+        Parameters
+        ----------
+        spatial : bool
+            Whether to collect additional stats.
+        bins : int or str
+            Number of bins or name of automatic algorithm of defining number of bins.
+        num_keep : int
+            Number of traces to store.
+        """
+        #pylint: disable=not-an-iterable
+        _ = kwargs
+
+        num_traces = len(self.segyfile.header)
+
+        # Get min/max values, store some of the traces
+        trace_container = []
+        value_min, value_max = np.inf, -np.inf
+
+        for i in tqdm(range(num_traces), desc='Finding min/max', ncols=1000, disable=(not pbar)):
+            trace = self.segyfile.trace[i]
+
+            trace_min, trace_max = find_min_max(trace)
+            if trace_min < value_min:
+                value_min = trace_min
+            if trace_max > value_max:
+                value_max = trace_max
+
+            if random() < (num_keep / num_traces) and trace_min != trace_max:
+                trace_container.extend(trace.tolist())
+                #TODO: add dtype for storing
+
+        # Collect more spatial stats: min, max, mean, std, histograms matrices
+        if spatial:
+            # Make bins
+            bins = np.histogram_bin_edges(None, bins, range=(value_min, value_max)).astype(np.float)
+            self.bins = bins
+
+            # Create containers
+            min_matrix, max_matrix = np.full(self.lens, np.nan), np.full(self.lens, np.nan)
+            hist_matrix = np.full((*self.lens, len(bins)-1), np.nan)
+
+            # Iterate over traces
+            description = f'Collecting stats for {self.name}'
+            for i in tqdm(range(num_traces), desc=description, ncols=1000, disable=(not pbar)):
+                trace = self.segyfile.trace[i]
+                header = self.segyfile.header[i]
+
+                # i -> id in a dataframe
+                keys = [header.get(field) for field in self.byte_no]
+                store_key = [self.uniques_inversed[j][item] for j, item in enumerate(keys)]
+                store_key = tuple(store_key)
+
+                # For each trace, we store an entire histogram of amplitudes
+                val_min, val_max = find_min_max(trace)
+                min_matrix[store_key] = val_min
+                max_matrix[store_key] = val_max
+
+                if val_min != val_max:
+                    histogram = np.histogram(trace, bins=bins)[0]
+                    hist_matrix[store_key] = histogram
+
+            # Restore stats from histogram
+            midpoints = (bins[1:] + bins[:-1]) / 2
+            probs = hist_matrix / np.sum(hist_matrix, axis=-1, keepdims=True)
+
+            mean_matrix = np.sum(probs * midpoints, axis=-1)
+            std_matrix = np.sqrt(np.sum((np.broadcast_to(midpoints, (*mean_matrix.shape, len(midpoints))) - \
+                                            mean_matrix.reshape(*mean_matrix.shape, 1))**2 * probs,
+                                        axis=-1))
+
+            # Store everything into instance
+            self.min_matrix, self.max_matrix = min_matrix, max_matrix
+            self.mean_matrix, self.std_matrix = mean_matrix, std_matrix
+            self.hist_matrix = hist_matrix
+            self.zero_traces = (min_matrix == max_matrix).astype(np.int)
+            self.zero_traces[np.isnan(min_matrix)] = 1
+
+        self.value_min, self.value_max = value_min, value_max
+        self.trace_container = np.array(trace_container)
+        self.q001, self.q01, self.q99, self.q999 = np.quantile(trace_container, [0.001, 0.01, 0.99, 0.999])
+        self.has_stats = True
+        self.store_meta()
+
+    def add_rotation_matrix(self):
+        """ Add transform from INLINE/CROSSLINE coordinates to CDP system. """
+        ix_points = []
+        cdp_points = []
+
+        for _ in range(3):
+            idx = np.random.randint(len(self.dataframe))
+            trace = self.segyfile.header[idx]
+
+            # INLINE_3D -> CDP_X, CROSSLINE_3D -> CDP_Y
+            ix = (trace[segyio.TraceField.INLINE_3D], trace[segyio.TraceField.CROSSLINE_3D])
+            cdp = (trace[segyio.TraceField.CDP_X], trace[segyio.TraceField.CDP_Y])
+
+            ix_points.append(ix)
+            cdp_points.append(cdp)
+        self.rotation_matrix = cv2.getAffineTransform(np.float32(ix_points), np.float32(cdp_points))
+
+    def lines_to_cdp(self, points):
+        """ Convert lines to CDP. """
+        return (self.rotation_matrix[:, :2] @ points.T + self.rotation_matrix[:, 2].reshape(2, -1)).T
+
+    def compute_area(self, correct=True, shift=50):
+        """ Compute approximate area of the cube in square kilometres.
+
+        Parameters
+        ----------
+        correct : bool
+            Whether to correct computed area for zero traces.
+        """
+        if self.headers != self.HEADERS_POST_FULL:
+            raise TypeError('Geometry index must be `POST_FULL`')
+
+        i = self.ilines[self.ilines_len // 2]
+        x = self.xlines[self.xlines_len // 2]
+
+        cdp_x, cdp_y = self.dataframe[['CDP_X', 'CDP_Y']].ix[(i, x)]
+        cdp_x_delta = abs(self.dataframe[['CDP_X']].ix[(i, x + shift)][0] - cdp_x)
+        cdp_y_delta = abs(self.dataframe[['CDP_Y']].ix[(i + shift, x)][0] - cdp_y)
+
+        if cdp_x_delta == 0 and cdp_y_delta == 0:
+            cdp_x_delta = abs(self.dataframe[['CDP_X']].ix[(i + shift, x)][0] - cdp_x)
+            cdp_y_delta = abs(self.dataframe[['CDP_Y']].ix[(i, x + shift)][0] - cdp_y)
+
+        cdp_x_delta /= shift
+        cdp_y_delta /= shift
+
+        ilines_km = cdp_y_delta * self.ilines_len / 1000
+        xlines_km = cdp_x_delta * self.xlines_len / 1000
+        area = ilines_km * xlines_km
+
+        if correct and hasattr(self, 'zero_traces'):
+            area -= (cdp_x_delta / 1000) * (cdp_y_delta / 1000) * np.sum(self.zero_traces)
+        return area
+
+
+    def set_index(self, index_headers, sortby=None):
+        """ Change current index to a subset of loaded headers. """
+        self.dataframe.reset_index(inplace=True)
+        if sortby:
+            self.dataframe.sort_values(index_headers, inplace=True, kind='mergesort')# the only stable sorting algorithm
+        self.dataframe.set_index(index_headers, inplace=True)
+        self.index_headers = index_headers
+        self.add_attributes()
+
+    # Methods to load actual data from SEG-Y
+    def load_trace(self, index):
+        """ Load individual trace from segyfile.
+        If passed `np.nan`, returns trace of zeros.
+        """
+        if not np.isnan(index):
+            return self.segyfile.trace.raw[int(index)]
+        return self._zero_trace
+
+    def load_traces(self, trace_indices):
+        """ Stack multiple traces together. """
+        return np.stack([self.load_trace(idx) for idx in trace_indices])
+
+
+    @lru_cache(128, attributes='index_headers')
+    def load_slide(self, loc=None, axis=0, start=None, end=None, step=1, stable=True):
+        """ Create indices and load actual traces for one slide.
+
+        If the current index is 1D, then slide is defined by `start`, `end`, `step`.
+        If the current index is 2D, then slide is defined by `loc` and `axis`.
+
+        Parameters
+        ----------
+        loc : int
+            Number of slide to load.
+        axis : int
+            Number of axis to load slide along.
+        start, end, step : ints
+            Parameters of slice loading for 1D index.
+        stable : bool
+            Whether or not to use the same sorting order as in the segyfile.
+        """
+        if axis in [0, 1]:
+            indices = self.make_slide_indices(loc=loc, start=start, end=end, step=step, axis=axis, stable=stable)
+            slide = self.load_traces(indices)
+        elif axis == 2:
+            slide = self.segyfile.depth_slice[loc].reshape(self.lens)
+        return slide
+
+    def make_slide_indices(self, loc=None, axis=0, start=None, end=None, step=1, stable=True, return_iterator=False):
+        """ Choose appropriate version of index creation for various lengths of current index.
+
+        Parameters
+        ----------
+        start, end, step : ints
+            Parameters of slice loading for 1d index.
+        stable : bool
+            Whether or not to use the same sorting order as in the segyfile.
+        return_iterator : bool
+            Whether to also return the same iterator that is used to index current `dataframe`.
+            Can be useful for subsequent loads from the same place in various instances.
+        """
+        if self.index_len == 1:
+            _ = loc, axis
+            result = self.make_slide_indices_1d(start=start, end=end, step=step, stable=stable,
+                                                return_iterator=return_iterator)
+        elif self.index_len == 2:
+            _ = start, end, step
+            result = self.make_slide_indices_2d(loc=loc, axis=axis, stable=stable,
+                                                return_iterator=return_iterator)
+        elif self.index_len == 3:
+            raise NotImplementedError('Yet to be done!')
+        else:
+            raise ValueError('Index lenght must be less than 4. ')
+        return result
+
+    def make_slide_indices_1d(self, start=None, end=None, step=1, stable=True, return_iterator=False):
+        """ 1D version of index creation. """
+        start = start or self.offsets[0]
+        end = end or self.uniques[0][-1]
+
+        if stable:
+            iterator = self.dataframe.index[(self.dataframe.index >= start) & (self.dataframe.index <= end)]
+            iterator = iterator.values[::step]
+        else:
+            iterator = np.arange(start, end+1, step)
+
+        indices = self.dataframe['trace_index'].reindex(iterator, fill_value=np.nan).values
+
+        if return_iterator:
+            return indices, iterator
+        return indices
+
+    def make_slide_indices_2d(self, loc, axis=0, stable=True, return_iterator=False):
+        """ 2D version of index creation. """
+        other_axis = 1 - axis
+        location = self.uniques[axis][loc]
+
+        if stable:
+            others = self.dataframe[self.dataframe.index.get_level_values(axis) == location]
+            others = others.index.get_level_values(other_axis).values
+        else:
+            others = self.uniques[other_axis]
+
+        iterator = list(zip([location] * len(others), others) if axis == 0 else zip(others, [location] * len(others)))
+        indices = self.dataframe['trace_index'].reindex(iterator, fill_value=np.nan).values
+
+        #TODO: keep only uniques, when needed, with `nan` filtering
+        if stable:
+            indices = np.unique(indices)
+
+        if return_iterator:
+            return indices, iterator
+        return indices
+
+
+    def _load_crop(self, locations):
+        """ Load 3D crop from the cube.
+
+        Parameters
+        ----------
+        locations : sequence of slices
+            List of desired slices to load: along the first index, the second, and depth.
+
+        Example
+        -------
+        If the current index is `INLINE_3D` and `CROSSLINE_3D`, then to load
+        5:110 ilines, 100:1105 crosslines, 0:700 depths, locations must be::
+            [slice(5, 110), slice(100, 1105), slice(0, 700)]
+        """
+        shape = np.array([((slc.stop or stop) - (slc.start or 0)) for slc, stop in zip(locations, self.cube_shape)])
+        indices = self.make_crop_indices(locations)
+        crop = self.load_traces(indices)[..., locations[-1]].reshape(shape)
+        return crop
+
+    def make_crop_indices(self, locations):
+        """ Create indices for 3D crop loading. """
+        iterator = list(product(*[[self.uniques[idx][i] for i in range(locations[idx].start, locations[idx].stop)]
+                                  for idx in range(2)]))
+        indices = self.dataframe['trace_index'].reindex(iterator, fill_value=np.nan).values
+        _, unique_ind = np.unique(indices, return_index=True)
+        return indices[np.sort(unique_ind, kind='stable')]
+
+    def load_crop(self, locations, threshold=15, mode='adaptive', **kwargs):
+        """ Smart choice between using :meth:`._load_crop` and stacking multiple slides created by :meth:`.load_slide`.
+
+        Parameters
+        ----------
+        mode : str
+            If `adaptive`, then function to load is chosen automatically.
+            If `slide` or `crop`, then uses that function to load data.
+        threshold : int
+            Upper bound for amount of slides to load. Used only in `adaptive` mode.
+        """
+        _ = kwargs
+        shape = np.array([((slc.stop or stop) - (slc.start or 0)) for slc, stop in zip(locations, self.cube_shape)])
+        axis = np.argmin(shape)
+        if mode == 'adaptive':
+            if axis in [0, 1]:
+                mode = 'slide' if min(shape) < threshold else 'crop'
+            else:
+                flag = np.prod(shape[:2]) / np.prod(self.cube_shape[:2])
+                mode = 'slide' if flag > 0.1 else 'crop'
+
+        if mode == 'slide':
+            slc = locations[axis]
+            if axis == 0:
+                return np.stack([self.load_slide(loc, axis=axis)[locations[1], locations[2]]
+                                 for loc in range(slc.start, slc.stop)], axis=axis)
+            if axis == 1:
+                return np.stack([self.load_slide(loc, axis=axis)[locations[0], locations[2]]
+                                 for loc in range(slc.start, slc.stop)], axis=axis)
+            if axis == 2:
+                return np.stack([self.load_slide(loc, axis=axis)[locations[0], locations[1]]
+                                 for loc in range(slc.start, slc.stop)], axis=axis)
+        return self._load_crop(locations)
+
+
+    def __getitem__(self, key):
+        """ Retrieve amplitudes from cube. Uses the usual `Numpy` semantics for indexing 3D array. """
+        key_ = list(key)
+        if len(key_) != len(self.cube_shape):
+            key_ += [slice(None)] * (len(self.cube_shape) - len(key_))
+
+        key, squeeze = [], []
+        for i, item in enumerate(key_):
+            max_size = self.cube_shape[i]
+
+            if isinstance(item, slice):
+                slc = slice(item.start or 0, item.stop or max_size)
+            elif isinstance(item, int):
+                item = item if item >= 0 else max_size - item
+                slc = slice(item, item + 1)
+                squeeze.append(i)
+            key.append(slc)
+
+        crop = self.load_crop(key)
+        if squeeze:
+            crop = np.squeeze(crop, axis=tuple(squeeze))
+        return crop
+
+    # Convert SEG-Y to HDF5
+    def make_hdf5(self, path_hdf5=None, postfix='', unsafe=True, store_meta=True, projections='ixh', pbar=True):
+        """ Converts `.segy` cube to `.hdf5` format.
+
+        Parameters
+        ----------
+        path_hdf5 : str
+            Path to store converted cube. By default, new cube is stored right next to original.
+        postfix : str
+            Postfix to add to the name of resulting cube.
+        """
+
+        cube_keys = {'i': 'cube', 'x': 'cube_x', 'h': 'cube_h'}
+        axes = {'i': [0, 1, 2], 'x': [1, 2, 0], 'h': [2, 0, 1]}
+
+        if self.index_headers != self.INDEX_POST and not unsafe:
+            # Currently supports only INLINE/CROSSLINE cubes
+            raise TypeError(f'Either set `unsafe=True` or set index to {self.INDEX_POST}')
+
+        path_hdf5 = path_hdf5 or (os.path.splitext(self.path)[0] + postfix + '.hdf5')
+
+        # Remove file, if exists: h5py can't do that
+        if os.path.exists(path_hdf5):
+            os.remove(path_hdf5)
+
+        # Create file and datasets inside
+        with h5py.File(path_hdf5, "w-") as file_hdf5:
+            cube_hdf5 = {
+                cube_keys[p]: file_hdf5.create_dataset(cube_keys[p], self.cube_shape[axes[p]]) for p in projections
+            }
+
+            # Default projection (ilines, xlines, depth) and depth-projection (depth, ilines, xlines)
+            total = 0
+            if 'i' in projections:
+                total += self.cube_shape[0]
+            if 'x' in projections:
+                total += self.cube_shape[1]
+            progress_bar = tqdm(total=total, ncols=1000, disable=(not pbar))
+
+            progress_bar.set_description(f'Converting {self.long_name}; ilines projection')
+            for i in range(self.cube_shape[0]):
+                slide = self.load_slide(i, stable=False)
+                if 'i' in projections:
+                    cube_hdf5['cube'][i, :, :] = slide.reshape(1, self.cube_shape[1], self.cube_shape[2])
+                if 'h' in projections:
+                    cube_hdf5['cube_h'][:, i, :] = slide.T
+                progress_bar.update()
+
+            # xline-oriented projection: (xlines, depth, ilines)
+            if 'x' in projections:
+                progress_bar.set_description(f'Converting {self.long_name} to hdf5; xlines projection')
+                for x in range(self.cube_shape[1]):
+                    slide = self.load_slide(x, axis=1, stable=False).T
+                    cube_hdf5['cube_x'][x, :, :,] = slide
+                    progress_bar.update()
+            progress_bar.close()
+
+        if store_meta:
+            if not self.has_stats:
+                self.collect_stats(pbar=pbar)
+
+            path_meta = os.path.splitext(path_hdf5)[0] + '.meta'
+            self.store_meta(path_meta)
+
+
+
+    # Convenient alias
+    convert_to_hdf5 = make_hdf5
+
+
